backend/requirements.txt
# Django Core
Django==5.0.3
djangorestframework==3.15.1
django-cors-headers==4.3.1
python-decouple==3.8

# Authentication
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# IBM Cloud Object Storage
ibm-cos-sdk==2.13.6
ibm-cos-sdk-core==2.13.6
ibm-cos-sdk-s3transfer==2.13.6
boto3==1.34.51

# Storage Backend
django-storages==1.14.2

# Image Processing
Pillow==10.2.0

# PDF Generation
reportlab==4.1.0

# Filtering and Search
django-filter==24.1

# WSGI Server (Production)
gunicorn==21.2.0

# Environment Variables
python-dotenv==1.0.1

# Utilities
pytz==2024.1

# HTTP Libraries (pinned for compatibility)
requests==2.32.2
urllib3==2.2.1
charset-normalizer==3.3.2
certifi==2024.2.2

# Made with Bob